+++ src/main/java/com/epam/aditechtask/service/ProductService.java
@@ -1,51 +1,78 @@
 package com.epam.aditechtask.service;
 
-import com.epam.aditechtask.dto.ProductDTO;
-import com.epam.aditechtask.persistence.entity.CategoryEntity;
+import com.epam.aditechtask.dto.ProductRequest;
+import com.epam.aditechtask.dto.ProductResponse;
+import com.epam.aditechtask.exception.ResourceNotFoundException;
+import com.epam.aditechtask.mapper.ProductMapper;
 import com.epam.aditechtask.persistence.entity.ProductEntity;
-import com.epam.aditechtask.persistence.repository.CategoryRepository;
 import com.epam.aditechtask.persistence.repository.ProductRepository;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@RequiredArgsConstructor
 public class ProductService {
 
+	private static final String DEFAULT_CURRENCY_CODE = "EUR";
+
 	private final ProductRepository productRepository;
-	private final CategoryRepository categoryRepository;
+	private final ExchangeRateService exchangeRateService;
+	private final ProductMapper productMapper;
 
-	public ProductService(ProductRepository productRepository, CategoryRepository categoryRepository) {
-		this.productRepository = productRepository;
-		this.categoryRepository = categoryRepository;
-	}
-
-	public List<ProductDTO> getAllProducts() {
-		return productRepository.findAll()
-			.stream()
-			.map(product -> new ProductDTO(product.getId(), product.getName(), product.getPriceEur(), product.getCategory().getId()))
+	public List<ProductResponse> getAllProducts(String currency) {
+		return productRepository.findAll().stream()
+			.map(product -> productMapper.toResponse(product, currency, getRateForCurrency(currency)))
 			.toList();
 	}
 
-	public ProductDTO getProductById(Long id) {
-		return productRepository.findById(id)
-			.map(product -> new ProductDTO(product.getId(), product.getName(), product.getPriceEur(), product.getCategory().getId()))
-			.orElseThrow(() -> new RuntimeException("Product not found"));
+	public ProductResponse getProductById(Long id, String currency) {
+		ProductEntity product = productRepository.findById(id)
+			.orElseThrow(() -> new ResourceNotFoundException("Product not found"));
+		return productMapper.toResponse(product, currency, getRateForCurrency(currency));
 	}
 
-	public ProductDTO createProduct(ProductDTO dto) {
-		CategoryEntity category = categoryRepository.findById(dto.categoryId())
-			.orElseThrow(() -> new RuntimeException("Category not found"));
+	public ProductResponse createProduct(ProductRequest request) {
+		BigDecimal rate = getRateForRequest(request);
+		ProductEntity product = productMapper.toEntity(request, rate);
+		return productMapper.toResponse(productRepository.save(product), request.currency().orElse(DEFAULT_CURRENCY_CODE), rate);
+	}
 
-		ProductEntity product = new ProductEntity();
-		product.setName(dto.name());
-		product.setPriceEur(dto.price());
-		product.setCategory(category);
-		product = productRepository.save(product);
+	@Transactional
+	public ProductResponse updateProduct(Long id, ProductRequest request) {
+		ProductEntity product = productRepository.findById(id)
+			.orElseThrow(() -> new ResourceNotFoundException("Product not found"));
 
-		return new ProductDTO(product.getId(), product.getName(), product.getPriceEur(), product.getCategory().getId());
+		BigDecimal rate = getRateForRequest(request);
+		product.setPriceEur(request.price().divide(rate, 2, RoundingMode.HALF_UP));
+
+
+		return productMapper.toResponse(productRepository.save(product), request.currency().orElse(DEFAULT_CURRENCY_CODE), rate);
 	}
 
 	public void deleteProduct(Long id) {
+		if (!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Product not found");
+		}
 		productRepository.deleteById(id);
+	}
+
+	private BigDecimal getRateForRequest(ProductRequest request) {
+		if (request.currency().isPresent() && !request.currency().get().equals(DEFAULT_CURRENCY_CODE)) {
+			return exchangeRateService.getExchangeRate(request.currency().get());
+		} else {
+			return BigDecimal.ONE;
+		}
+	}
+
+	private BigDecimal getRateForCurrency(String currency) {
+		if (currency.equals(DEFAULT_CURRENCY_CODE)) {
+			return BigDecimal.ONE;
+		} else {
+			return exchangeRateService.getExchangeRate(currency);
+		}
 	}
 }
